+++ Packages/com.unity.inputsystem/InputSystem/NativeInputRuntime.cs
@@ -215,12 +215,7 @@ namespace UnityEngine.Experimental.Input.LowLevel
         public double currentTimeForFixedUpdate => Time.fixedUnscaledTime + currentTimeOffsetToRealtimeSinceStartup;
 
         public double currentTimeOffsetToRealtimeSinceStartup => NativeInputSystem.currentTimeOffsetToRealtimeSinceStartup;
-
-        public bool shouldRunInBackground
-        {
-            set => NativeInputSystem.SetUpdateMask((NativeInputUpdateType)(value ? 1 << 31 : 0));
-        }
-
+        
         private Action m_ShutdownMethod;
 
         private void OnShutdown()